chore: concert / ticketing network code 분리
