+++ Boolti/Boolti/Sources/Network/DTO/Ticketing/Request/CheckInvitationCodeRequestDTO.swift
