+++ Boolti/Boolti/Sources/Network/DTO/Ticketing/Request/InvitationTicketingRequestDTO.swift
