+++ Boolti/Boolti/Sources/Network/DTO/Ticketing/Response/CheckInvitationCodeResponseDTO.swift
