+++ Boolti/Boolti/Sources/Network/DTO/Ticketing/Response/SalesTicketResponseDTO.swift
